Updated to libg++ 2.4
Former-commit-id: c21270b66cc965eb2ea6f816046c224c450243a0

+++ gnu/lib/libg++/libg++/Fix16.h
@@ -127,7 +127,7 @@ begin_define
 define|#
 directive|define
 name|Fix32_m_max
-value|((1<< 31) - 1)
+value|((long)((1<< 31) - 1))
 end_define
 
 begin_define
